Use the Webpack dependency graph throughout
This is obviously a very non-essential change for a small project like
this. Nevertheless, it shows a powerful aspect of Webpack, namely that
it can help you take full control over the project's dependencies, be
they Javascript modules, CSS files or assets like images or the `.tsv`
file in use here.

The reason this is more powerful than using the CopyWebpackPlugin is
that it allows you to fingerprint files or move them around when they
are built, e.g. onto a CDN. In order to make even fuller use of this,
you can make use of the HtmlWebpackPlugin [1] to generate the HTML
needed. This allows you to also fingerprint the JS assets, etc. But I
digress, this is too much for this project :)

    new HtmlWebpackPlugin({
      template: 'html/index.html',
      inject: 'body'
    })

[1] https://github.com/ampedandwired/html-webpack-plugin