Optimize production builds by passing in NODE_ENV
React contains a lot of debugging machinery that is being run throughout
development. This should always be stripped out in production in order
to make the application more performant. The most simple way to do this
is with the built-in Webpack EnvironmentPlugin that will make the
environment variables given available to the script and make React strip
out debugging calls when NODE_ENV is not set to “development”.

There is a more advanced way of achieving the same with the
DefinePlugin, but this power is not needed here. The DefinePlugin has
the added benefit of allowing you to define your own “globals” and make
use of them in the script.

    new webpack.DefinePlugin({
      '__DEV__': JSON.stringify(false),
      '__MY_OTHER_GLOBAL__': 'A nice value',
      'process.env.NODE_ENV': JSON.stringify('production')
    })